fix code metrics problems
