+++ lib/ui/screen/sight_details.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:places/domain/sight.dart';
+import 'package:places/ui/widgets/icon_text_button.dart';
 
 class SightDetails extends StatelessWidget {
   final Sight sight;
@@ -9,6 +10,7 @@ class SightDetails extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     double childMargin = 24;
+
     return Scaffold(
       body: Column(
         children: [
@@ -104,37 +106,6 @@ class SightDetails extends StatelessWidget {
           ),
         ],
       ),
-    );
-  }
-}
-
-class IconTextButton extends StatelessWidget {
-  final IconData icon;
-  final String name;
-  final bool isActive;
-  const IconTextButton({
-    required this.icon,
-    required this.name,
-    required this.isActive,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    Color color = isActive ? Color(0xff3B3E5B) : Color.fromRGBO(124, 126, 146, 0.56);
-    return Row(
-      children: [
-        Icon(
-          icon,
-          color: color,
-        ),
-        Padding(
-          padding: const EdgeInsets.only(left: 9.0),
-          child: Text(
-            name,
-            style: TextStyle(fontSize: 14, fontWeight: FontWeight.w400, color: color),
-          ),
-        )
-      ],
     );
   }
 }
