+++ test/widget_test.dart
@@ -5,10 +5,10 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
+// import 'package:flutter/material.dart';
+// import 'package:flutter_test/flutter_test.dart';
 
-import 'package:places/main.dart';
+// import 'package:places/main.dart';
 
 void main() {
   // testWidgets('App find hello test', (WidgetTester tester) async {
